加入显示天气信息的功能

+++ app/src/main/java/com/example/ativ/weather_forecast/gson/Forecast.java
@@ -18,7 +18,7 @@ public class Forecast {
     @SerializedName("sunset")
     public String sunset;
     @SerializedName("aqi")
-    public int aqi;
+    public float aqi;
     @SerializedName("fx")
     public String fx;
     @SerializedName("fl")
